PagingType move to common type
