+++ src/model/usecase/company/select.tsx
@@ -3,8 +3,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import { TableType } from '~/components';
 import { useCompanyRepo } from '~/model/repository';
 import { Company } from '~/model/entity';
-import { useRouter } from 'next/router';
-import { PagingType } from '~/model/repository/use-company';
+import { PagingType } from '~/types/common';
 
 export const useCompanySelectUseCase = () => {
   const [companyList, setCompanyList] = useState<Company[]>([]);
